fix for Opera
